backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
openai==1.58.1
pdfplumber==0.11.4
python-docx==1.1.2
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
supabase==2.11.0
python-jose[cryptography]==3.3.0
httpx==0.28.1
